ISSUE #4909 - move calibration button together with other toolbar buttons
